backend/requirements.txt
anthropic
python-dotenv
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
PyPDF2
python-docx
